Move findParentWithClass outside psi subdir
